+++ src/config/mod.rs
@@ -9,11 +9,10 @@ pub use self::range::Range;
 pub use self::config::{Config, read_config, find_config};
 
 pub mod settings;
+pub mod read_settings;
 pub mod containers;
 pub mod range;
 pub mod builders;
 pub mod config;
 pub mod command;
 pub mod validate;
-
-