+++ src/config/settings.rs
@@ -7,6 +7,7 @@ use rustc_serialize::json;
 #[derive(RustcDecodable, RustcEncodable, Default, Clone)]
 pub struct Settings {
     pub version_check: bool,
+    pub proxy_env_vars: bool,
     pub ubuntu_mirror: String,
     pub alpine_mirror: Option<String>,
     pub uid_map: Option<(Vec<(uid_t, uid_t, uid_t)>,