+++ src/wrapper/commandline.rs
@@ -65,7 +65,7 @@ pub fn commandline_cmd(command: &CommandInfo,
     let cont_ver = wrapper.root.as_ref().unwrap();
     try!(setup::setup_filesystem(cconfig, write_mode, &cont_ver));
 
-    let mut env = try!(setup::get_environment(cconfig));
+    let mut env = try!(setup::get_environment(cconfig, &wrapper.settings));
     for (k, v) in command.environ.iter() {
         env.insert(k.clone(), v.clone());
     }